Move Yen tests into their own file
